Add files via upload
I edited the database labels, and text boxes. I also changed the type on two of them from the vendor table to labels. I coded the exit button and the form_Closing procedure. I didnt re add the menu bar items. They are currently missing (querry and exit). I added the options to the top and coded the mainformload. I begun coding the movenext button, however, this is where I have run into an error. No point in coding the others till this is fixed, because the code is the same across them all. I am currently waiting for assistance to resolve this issue. If you get a chance to look at it, let me know if you find something that works. - By Corita